Apply Hexagonal Architecture

+++ src/main/java/org/puzre/core/port/service/ITypeService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface ITypeService {
 
     List<Type> listAllTypes();
-    Type findTypeById(int typeId);
+    Type findTypeById(Long typeId);
 
 }
